# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ATHAR API",
    "version": "2.0.0a0"
  },
  "paths": {
...
    "/runs": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "List Runs",
        "operationId": "list_runs_runs_get",
        "parameters": [
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunSummary"
                  },
                  "title": "Response List Runs Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Run",
        "operationId": "get_run_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunManifest"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/events": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "List Events",
        "operationId": "list_events_runs__run_id__events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1000,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Events Runs  Run Id  Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/events/stream": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Stream Events",
        "operationId": "stream_events_runs__run_id__events_stream_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/report": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Report",
        "operationId": "get_report_runs__run_id__report_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Report Runs  Run Id  Report Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/report.html": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Export Report Html",
        "description": "Chromium-free preview of the exact document the PDF prints.",
        "operationId": "export_report_html_runs__run_id__report_html_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "locale",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "ar",
              "title": "Locale"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/report.pdf": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Export Report Pdf",
        "operationId": "export_report_pdf_runs__run_id__report_pdf_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "locale",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "ar",
              "title": "Locale"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/timeline": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Timeline",
        "description": "Cross-camera timeline: per-camera scene-clock coverage + identity\nspans, derived from package.report and the run's time base.",
        "operationId": "get_timeline_runs__run_id__timeline_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/thumbs/{camera_id}/{track_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Thumbnail",
        "operationId": "get_thumbnail_runs__run_id__thumbs__camera_id___track_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "camera_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Camera Id"
            }
          },
          {
            "name": "track_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Track Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/clips/{camera_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Clip",
        "description": "Evidence clip for a scene-clock span of one camera, transcoded on\ndemand (cached under the run dir).",
        "operationId": "get_clip_runs__run_id__clips__camera_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "camera_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Camera Id"
            }
          },
          {
            "name": "start_s",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "Start S"
            }
          },
          {
            "name": "end_s",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "End S"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/artifacts/{name}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Download Artifact",
        "operationId": "download_artifact_runs__run_id__artifacts__name__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/cases/{case_id}/runs": {
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "Attach Run",
        "operationId": "attach_run_cases__case_id__runs_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachRunRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cases/{case_id}/runs/{run_id}": {
      "delete": {
        "tags": [
          "cases"
        ],
        "summary": "Detach Run",
        "operationId": "detach_run_cases__case_id__runs__run_id__delete",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttachRunRequest": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          }
        },
        "type": "object",
        "required": [
          "run_id"
        ],
        "title": "AttachRunRequest"
      },
...
      "CaseDetail": {
        "properties": {
          "case_id": {
            "type": "string",
            "title": "Case Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "owner": {
            "type": "string",
            "title": "Owner"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/CaseRunOut"
            },
            "type": "array",
            "title": "Runs"
          },
          "targets": {
            "items": {
              "$ref": "#/components/schemas/TargetOut"
            },
            "type": "array",
            "title": "Targets"
          }
        },
        "type": "object",
        "required": [
          "case_id",
          "title",
          "status",
          "owner",
          "created_at",
          "updated_at",
          "runs",
          "targets"
        ],
        "title": "CaseDetail"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RunManifest": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "title": "Schema Version",
            "default": 1
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "role": {
            "$ref": "#/components/schemas/RunRole"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus",
            "default": "created"
          },
          "profile_name": {
            "type": "string",
            "title": "Profile Name"
          },
          "case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By",
            "description": "Operator identity (audit)"
          },
          "inputs": {
            "items": {
              "$ref": "#/components/schemas/VideoInput"
            },
            "type": "array",
            "title": "Inputs"
          },
          "timebase": {
            "$ref": "#/components/schemas/SceneClock"
          },
          "config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResolvedConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "artifacts": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ArtifactRecord"
            },
            "type": "object",
            "title": "Artifacts"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "role",
          "profile_name"
        ],
        "title": "RunManifest",
        "description": "Complete, self-describing record of one pipeline run."
      },
      "RunRole": {
        "type": "string",
        "enum": [
          "gallery",
          "probe",
          "benchmark",
          "adaptation"
        ],
        "title": "RunRole"
      },
...
      "RunSummary": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "profile_name": {
            "type": "string",
            "title": "Profile Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "config_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Hash"
          },
          "num_artifacts": {
            "type": "integer",
            "title": "Num Artifacts"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "role",
          "status",
          "profile_name",
          "created_at",
          "config_hash",
          "num_artifacts",
          "error"
        ],
        "title": "RunSummary"
      },
...
      "TimelineOut": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "span_start_s": {
            "type": "number",
            "title": "Span Start S"
          },
          "span_end_s": {
            "type": "number",
            "title": "Span End S"
          },
          "cameras": {
            "items": {
              "$ref": "#/components/schemas/TimelineCameraOut"
            },
            "type": "array",
            "title": "Cameras"
          },
          "identities": {
            "items": {
              "$ref": "#/components/schemas/TimelineIdentityOut"
            },
            "type": "array",
            "title": "Identities"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "span_start_s",
          "span_end_s",
          "cameras",
          "identities"
        ],
        "title": "TimelineOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}